+++ example/example.dart
@@ -2,14 +2,11 @@ import 'package:flutter_simple_dependency_injection/injector.dart';
 
 void main() {
   final injector = Injector.getInjector();
-  injector.map(Logger, (i) => new Logger(), isSingleton: true);
-  injector.map(String, (i) => "https://api.com/", key: "apiUrl");
-  injector.map(SomeService,
-      (i) => new SomeService(i.get(Logger), i.get(String, "apiUrl")));
+  injector.map<Logger>((i) => new Logger(), isSingleton: true);
+  injector.map<String>((i) => "https://api.com/", key: "apiUrl");
+  injector.map<SomeService>((i) => new SomeService(i.get<Logger>(), i.get<String>("apiUrl")));
 
-  injector.get<SomeService>(SomeService).doSomething();
-  // passing in the [SomeService] as a generic parameter strongly types the return object.
-  injector.get<SomeService>(SomeService).doSomething();
+  injector.get<SomeService>().doSomething();
 }
 
 class Logger {