+++ test/flutter_simple_dependency_injection_test.dart
@@ -37,12 +37,11 @@ void main() {
   test("can have isolated injector instances", () async {
     final injectorOne = Injector.getInjector("one");
     final injectorTwo = Injector.getInjector("two");
-    injectorOne.map(
-        ObjectWithNoDependencies, (i) => new ObjectWithNoDependencies());
-    final instance = injectorOne.get(ObjectWithNoDependencies);
+    injectorOne.map<ObjectWithNoDependencies>((i) => new ObjectWithNoDependencies());
+    final instance = injectorOne.get<ObjectWithNoDependencies>();
     expect(instance is ObjectWithNoDependencies, true);
     expect(
-        () => injectorTwo.get(ObjectWithNoDependencies),
+        () => injectorTwo.get<ObjectWithNoDependencies>(),
         throwsA(predicate((e) =>
             e is InjectorException &&
             e.message ==
@@ -51,10 +50,9 @@ void main() {
 
   test("can map class factory and get object instance", () async {
     final injector = Injector.getInjector();
-    injector.map(
-        ObjectWithNoDependencies, (injector) => new ObjectWithNoDependencies());
+    injector.map<ObjectWithNoDependencies>((injector) => new ObjectWithNoDependencies());
     final instance =
-        injector.get<ObjectWithNoDependencies>(ObjectWithNoDependencies);
+        injector.get<ObjectWithNoDependencies>();
     expect(instance is ObjectWithNoDependencies, true);
     expect(instance.propertyOne, "Hello");
     injector.dispose();
@@ -64,12 +62,11 @@ void main() {
       "can map class factory for singleton and always get back the same object",
       () async {
     final injector = Injector.getInjector();
-    injector.map(
-        ObjectWithNoDependencies, (injector) => new ObjectWithNoDependencies(),
+    injector.map<ObjectWithNoDependencies>((injector) => new ObjectWithNoDependencies(),
         isSingleton: true);
-    final instanceOne = injector.get(ObjectWithNoDependencies);
-    final instanceTwo = injector.get(ObjectWithNoDependencies);
-    final instanceThree = injector.get(ObjectWithNoDependencies);
+    final instanceOne = injector.get<ObjectWithNoDependencies>();
+    final instanceTwo = injector.get<ObjectWithNoDependencies>();
+    final instanceThree = injector.get<ObjectWithNoDependencies>();
     expect(instanceOne is ObjectWithNoDependencies, true);
     expect(instanceTwo.hashCode, instanceOne.hashCode);
     expect(instanceThree.hashCode, instanceOne.hashCode);
@@ -78,13 +75,10 @@ void main() {
 
   test("can map class factory and get object simple instance", () async {
     final injector = Injector.getInjector();
-    injector.map(
-        ObjectWithNoDependencies, (injector) => new ObjectWithNoDependencies());
-    injector.map(
-        ObjectWithOneDependency,
-        (injector) => new ObjectWithOneDependency(
-            injector.get(ObjectWithNoDependencies)));
-    final instance = injector.get(ObjectWithOneDependency);
+    injector.map<ObjectWithNoDependencies>((injector) => new ObjectWithNoDependencies());
+    injector.map<ObjectWithOneDependency>((injector) => new ObjectWithOneDependency(
+            injector.get<ObjectWithNoDependencies>()));
+    final instance = injector.get<ObjectWithOneDependency>();
     expect(instance is ObjectWithOneDependency, true);
     expect(instance.dependencyOne is ObjectWithNoDependencies, true);
     expect(instance.propertyOne, "Hello Jon");
@@ -93,13 +87,10 @@ void main() {
 
   test("can map same object with different keys and get instances", () async {
     final injector = Injector.getInjector();
-    injector.map(
-        ObjectWithNoDependencies, (injector) => new ObjectWithNoDependencies());
-    injector.map(
-        ObjectWithNoDependencies, (injector) => new ObjectWithNoDependencies(),
-        key: "One");
-    final instanceOne = injector.get(ObjectWithNoDependencies);
-    final instanceTwo = injector.get(ObjectWithNoDependencies, "One");
+    injector.map<ObjectWithNoDependencies>((injector) => new ObjectWithNoDependencies());
+    injector.map<ObjectWithNoDependencies>((injector) => new ObjectWithNoDependencies(), key: "One");
+    final instanceOne = injector.get<ObjectWithNoDependencies>();
+    final instanceTwo = injector.get<ObjectWithNoDependencies>("One");
     expect(instanceOne is ObjectWithNoDependencies, true);
     expect(instanceTwo is ObjectWithNoDependencies, true);
     expect(instanceOne != instanceTwo, true);
@@ -108,9 +99,9 @@ void main() {
 
   test("can map simple named string type", () async {
     final injector = Injector.getInjector();
-    injector.map(String, (injector) => "Jon", key: "MyName");
-    final instanceOne = injector.get(String, "MyName");
-    final instanceTwo = injector.get(String, "MyName");
+    injector.map<String>((injector) => "Jon", key: "MyName");
+    final instanceOne = injector.get<String>("MyName");
+    final instanceTwo = injector.get<String>("MyName");
     expect(instanceOne is String, true);
     expect(instanceOne, "Jon");
     expect(instanceTwo, instanceOne);
@@ -119,12 +110,11 @@ void main() {
 
   test("exception thrown when type is not known", () async {
     final injector = Injector.getInjector();
-    injector.map(
-        ObjectWithOneDependency,
+    injector.map<ObjectWithOneDependency>(
         (injector) => new ObjectWithOneDependency(
-            injector.get(ObjectWithNoDependencies)));
+            injector.get<ObjectWithNoDependencies>()));
     expect(
-        () => injector.get(ObjectWithOneDependency),
+        () => injector.get<ObjectWithOneDependency>(),
         throwsA(predicate((e) =>
             e is InjectorException &&
             e.message ==
@@ -134,12 +124,11 @@ void main() {
 
   test("exception thrown when keyed type is not known", () async {
     final injector = Injector.getInjector();
-    injector.map(
-        ObjectWithOneDependency,
+    injector.map<ObjectWithOneDependency>(
         (injector) => new ObjectWithOneDependency(
-            injector.get(ObjectWithNoDependencies, "Key")));
+            injector.get<ObjectWithNoDependencies>("Key")));
     expect(
-        () => injector.get(ObjectWithOneDependency),
+        () => injector.get<ObjectWithOneDependency>(),
         throwsA(predicate((e) =>
             e is InjectorException &&
             e.message ==